refactor firebase gateway to services module
